add liquibase plugin to deploy sql changes fix bug with attachment fix bug with creation new nomenclature from order screen

+++ src/main/java/com/app/screen/TestScreen.java
@@ -125,7 +125,7 @@ public class TestScreen {
         ExternalContext ec = fc.getExternalContext();
         ec.responseReset();
         ec.setResponseContentType(attachment.getType());
-        ec.setResponseContentLength((int) attachment.getSize());
+        ec.setResponseContentLength(Math.toIntExact(attachment.getSize()));
         ec.setResponseHeader("Content-Disposition", "attachment; filename=\"" + attachment.getName() + "\"");
 
         try {
